requirements.txt
instagrapi==2.0.0
python-dotenv==1.0.0
schedule==1.2.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
apscheduler==3.10.4 